refactor(desktop): extract main process IPC handlers
